+++ src/ansys/simai/core/client.py
@@ -35,8 +35,9 @@ from ansys.simai.core.data.global_coefficients_requests import (
 )
 from ansys.simai.core.data.models import ModelDirectory
 from ansys.simai.core.data.optimizations import (
+    LegacyOptimizationDirectory,
     OptimizationDirectory,
-    _OptimizationTrialRunDirectory,
+    _LegacyOptimizationTrialRunDirectory,
 )
 from ansys.simai.core.data.post_processings import PostProcessingDirectory
 from ansys.simai.core.data.predictions import PredictionDirectory
@@ -88,7 +89,8 @@ class SimAIClient:
         self._process_gc_formula_directory = ProcessGlobalCoefficientDirectory(client=self)
         self._geometry_directory = GeometryDirectory(client=self)
         self._optimization_directory = OptimizationDirectory(client=self)
-        self._optimization_trial_run_directory = _OptimizationTrialRunDirectory(client=self)
+        self._legacy_optimization_directory = LegacyOptimizationDirectory(client=self)
+        self._optimization_trial_run_directory = _LegacyOptimizationTrialRunDirectory(client=self)
         self._post_processing_directory = PostProcessingDirectory(client=self)
         self._project_directory = ProjectDirectory(client=self)
         self._model_directory = ModelDirectory(client=self)
@@ -210,6 +212,13 @@ class SimAIClient:
         For more information, see :ref:`optimizations`.
         """
         return self._optimization_directory
+
+    @property
+    def legacy_optimizations(self):
+        """Representation of all legacy client side optimizations
+        For more information, see :ref:`optimizations`.
+        """
+        return self._legacy_optimization_directory
 
     @property
     def training_data(self):